QuickStore/pyproject.toml
[project]
name = "quickstore"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djoser>=2.3.3",
    "drf-nested-routers>=0.95.0",
    "mysqlclient>=2.2.7",
]
